% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp20028.R
\name{Prim}
\alias{Prim}
\title{Prim  Algorithm}
\usage{
Prim(vn, adjacency_matrix, myl)
}
\arguments{
\item{vn}{The number of vertices in the graph (numeric)}

\item{adjacency_matrix}{the adjacency matrix of the graph (matrix)}

\item{myl}{the layout of the graph (matrix)}
}
\value{
the MST of the graph
}
\description{
Prim  Algorithm
}
\examples{
\dontrun{
vn <- 12
a <- c(0, 2, 0, 0, 3, 0, 0, 0, 0, 0, 0, 0,
    2, 0, 3, 0, 0, 1, 0, 0, 0, 0, 0, 0,
    0, 3, 0, 1, 0, 0, 2, 0, 0, 0, 0, 0,
    0, 0, 1, 0, 0, 0, 0, 2, 0, 0, 0, 0,
    3, 0, 0, 0, 0, 4, 0, 0, 4, 0, 0, 0,
    0, 1, 0, 0, 4, 0, 3, 0, 0, 2, 0, 0,
    0, 0, 2, 0, 0, 3, 0, 3, 0, 0, 4, 3,
    0, 0, 0, 2, 0, 0, 3, 0, 0, 0, 0, 0,
    0, 0, 0, 0, 4, 0, 0, 0, 0, 3, 0, 0,
    0, 0, 0, 0, 0, 2, 0, 0, 3, 0, 3, 0,
    0, 0, 0, 0, 0, 0, 4, 0, 0, 3, 0, 1,
    0, 0, 0, 0, 0, 0, 3, 0, 0, 0, 1, 0)
adjacency_matrix <- matrix(data = a, nrow = 12, ncol = 12, dimnames=list(letters[1:vn],letters[1:vn]))
b <- c(0,1,2,3,0,1,2,3,0,1,2,3,2,2,2,2,1,1,1,1,0,0,0,0)
myl <- matrix(data = b, nrow = 12, ncol = 2)
rm(a)
res <- Prim(vn, adjacency_matrix, myl)
}
}
